agent/requirements.txt
# Agent Framework dependencies (pinned versions)
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16
agent-dev-cli==0.0.1b260316
agent-framework-core==1.0.0rc6
agent-framework-foundry==1.0.0rc6
agent-framework-openai==1.0.0rc6

# Azure identity
azure-identity

# Environment management
python-dotenv

# HTTP server
uvicorn[standard]
fastapi

# Database (for accessing ERP data)
psycopg2-binary

# Tracing and evaluation
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
azure-monitor-opentelemetry

# Evaluation
azure-ai-evaluation